# farin-main/docs/swagger.yaml
definitions:
...
  farin_app_api_response.Response-farin_domain_dto_DriverAssignmentListResponse:
    properties:
      data:
        $ref: '#/definitions/farin_domain_dto.DriverAssignmentListResponse'
      message:
        type: string
      status:
        type: integer
    type: object
  farin_app_api_response.Response-farin_domain_dto_DriverAssignmentResponse:
    properties:
      data:
        $ref: '#/definitions/farin_domain_dto.DriverAssignmentResponse'
      message:
        type: string
      status:
        type: integer
    type: object
...
  farin_app_api_response.Response-swagger_EmptyObject:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyObject'
      message:
        type: string
      status:
        type: integer
    type: object
...
  farin_domain_dto.DriverAssignmentListResponse:
    properties:
      driverAssignments:
        items:
          $ref: '#/definitions/farin_domain_dto.DriverAssignmentResponse'
        type: array
      total:
        type: integer
    type: object
  farin_domain_dto.DriverAssignmentRequest:
    properties:
      calenderID:
        type: string
      codeVehicle:
        type: string
      driverId:
        type: string
      ringId:
        type: integer
    required:
    - calenderID
    - codeVehicle
    - driverId
    - ringId
    type: object
  farin_domain_dto.DriverAssignmentResponse:
    properties:
      calendarId:
        type: string
      codeVehicle:
        type: string
      createdAt:
        type: integer
      driverId:
        type: string
      id:
        type: string
      ringId:
        type: integer
      updatedAt:
        type: integer
    type: object
...
  swagger.EmptyObject:
    type: object
info:
  contact: {}
paths:
...
  /driver-assignments:
    get:
      consumes:
      - application/json
      description: Retrieve a list of driverAssignments with optional filters and
        pagination
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - default: created_at
        description: Field to sort by
        in: query
        name: sortField
        type: string
      - default: asc
        description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_DriverAssignmentListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: List driverAssignments
      tags:
      - driverAssignments
    post:
      consumes:
      - application/json
      description: Create a new driverAssignment by providing driverAssignment details
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: DriverAssignment Data
        in: body
        name: driverAssignment
        required: true
        schema:
          $ref: '#/definitions/farin_domain_dto.DriverAssignmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_DriverAssignmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Create a new driverAssignment
      tags:
      - driverAssignments
  /driver-assignments/{id}:
    delete:
      description: Delete a driverAssignment by ID
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: DriverAssignment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Delete a driverAssignment
      tags:
      - driverAssignments
    get:
      description: Retrieve driverAssignment details by ID
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: DriverAssignment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_DriverAssignmentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Get driverAssignment details
      tags:
      - driverAssignments
    put:
      consumes:
      - application/json
      description: Update an existing driverAssignment's details
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated DriverAssignment Data
        in: body
        name: driverAssignment
        required: true
        schema:
          $ref: '#/definitions/farin_domain_dto.DriverAssignmentRequest'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_DriverAssignmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Update driverAssignment details
      tags:
      - driverAssignments
...
swagger: "2.0"